refactor(aviation): extract AviationClient HTTP into client.rs
